// frontend/src/app/utils/enums.ts
export enum AppRoute {
  Auth = 'auth',
  Admin = 'admin',
  Order = 'order',
  Orders = 'orders',
  Suborders = 'suborders',
  Dashboard = 'dashboard'
}

export enum AppPath {
  Auth = '/auth',
  Admin = '/admin',
  Order = '/order',
  Suborders = '/suborders',
  Dashboard = '/dashboard',
  AllOrders = '/orders/all',
  UserOrders = '/orders/user'
}